Record and record pattern

+++ Chapter04/P89_RecordSerialization/src/main/java/modern/challenge/MelonRecord.java
@@ -3,7 +3,7 @@ package modern.challenge;
 import java.io.Serializable;
 import java.util.Objects;
 
-public record MelonRecord(String type, int weight) implements Serializable {
+public record MelonRecord(String type, float weight) implements Serializable {
     
     public MelonRecord {
         Objects.requireNonNull(type, "The melon's type cannot be null");
